refactor(dashboards): TeacherDashboard sous 300 lignes — éclatement complet (H11)
Vue 692 → orchestrateur 155 l. Logique extraite dans useTeacherDashboard.js
(dashboard KLASSCI + cache, navigation matière, formatDate). Template scindé
en 6 composants : DashboardStatCards, DashboardActivityWidgets,
DashboardMatieresList (émet navigate), DashboardClassesList,
DashboardEvaluationsList (formatDate en prop) et DashboardQuickActions.
CSS déplacé verbatim (chrome widget/stat dupliqué par composant scopé,
@media stat-value conservé dans les deux composants concernés pour parité).
Parité : mêmes appels services/cache, mêmes routes, mêmes classes/textes.
Tests : composable (6) + 6 montages.
